+++ packages/winnow/lib/prepare-sql.js
@@ -1,8 +1,8 @@
-const alasql = require('./alasql')
+const filterAndTransform = require('./filter-and-transform')
 const Params = require('./params')
 
 module.exports = function (statement) {
-  const query = alasql.compile(statement)
+  const query = filterAndTransform.compile(statement)
   return function (inParams) {
     const params = Params.prepare(inParams)
     const results = query(params)