+++ packages/winnow/test/unit/filter-and-transform/index.spec.js
@@ -1,5 +1,5 @@
 const test = require('tape')
-const sql = require('../../lib/alasql')
+const filterAndTransform = require('../../../lib/filter-and-transform')
 const polygonFilter = {
   type: 'Polygon',
   coordinates: [
@@ -73,16 +73,18 @@ const polygonFeature = {
 
 test('sql.fn.ST_Within - geometries fully within a target polygon should return true', t => {
   t.plan(3)
-  t.ok(sql.fn.ST_Within(pointFeature, polygonFilter), 'point within filter geom')
-  t.ok(sql.fn.ST_Within(lineFeature, polygonFilter), 'line within filter geom')
-  t.ok(sql.fn.ST_Within(polygonFeature, polygonFilter), 'polygon within filter geom')
+  t.ok(filterAndTransform.fn.ST_Within(pointFeature, polygonFilter), 'point within filter geom')
+  t.ok(filterAndTransform.fn.ST_Within(lineFeature, polygonFilter), 'line within filter geom')
+  t.ok(filterAndTransform.fn.ST_Within(polygonFeature, polygonFilter), 'polygon within filter geom')
 })
 
 test('sql.fn.ST_Within - falsey feature geometries should return false', t => {
   const falseyFeatures = [undefined, null, {}, { coordinates: [] }, { coordinates: [0, 0] }, { type: 'Point' }]
 
   falseyFeatures.forEach(falsey => {
-    t.notOk(sql.fn.ST_Within(falsey, polygonFilter), 'falsey feature should return false')
+    t.notOk(filterAndTransform.fn.ST_Within(falsey, polygonFilter), 'falsey feature should return false')
   })
   t.end()
 })
+
+// TODO: put other method decorators under test